+++ cmake/copy_and_remove_0.ps1
@@ -1,4 +1,4 @@
-robocopy "$($args[0])" "$($args[1])" /s
+robocopy "$($args[0])" "$($args[1])" /s /nfl /ndl /nc /ns /np
 Get-ChildItem "$($args[1])" -recurse | where-object {$_.length -eq 0} | ?{Remove-Item $_.fullname}
 Get-ChildItem "$($args[1])" -recurse | where-object {$_.Name -match "_docify"} | ?{Remove-Item $_.fullname -recurse}
 Get-ChildItem "$($args[1])" -recurse | where-object {($_.Name -match "@*_old") -and ($_.GetType().Name -eq "DirectoryInfo")} | ?{Remove-Item $_.fullname -recurse}